LoyolaHouse/requirements.txt
Django>=5.1.7
mysqlclient>=2.1.0
django-ckeditor-5>=0.2.10
twilio>=8.0.0
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
